requirements.txt
# Friday Reports Automation - Dependencies

# Google API clients
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-analytics-data>=0.18.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel file support

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# OpenAI API
openai>=1.0.0

# Email processing
premailer>=3.10.0  # CSS inlining for emails

# SFTP uploads
paramiko>=3.3.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
